Crate pulldown_cmark_to_cmark

Structs

Heading

ImageLink

Options

Configuration for the cmark_with_options() and cmark_resume_with_options() functions. The defaults should provide decent spacing and most importantly, will provide a faithful rendering of your markdown document particularly when rendering it to HTML.

State

The state of the cmark_resume() and cmark_resume_with_options() functions. This does not only allow introspection, but enables the user to halt the serialization at any time, and resume it later.

Enums

Alignment

Similar to Pulldown-Cmark-Alignment, but with required traits for comparison to allow testing.

LinkCategory

Constants

DEFAULT_CODE_BLOCK_TOKEN_COUNT

Thea mount of code-block tokens one needs to produce a valid fenced code-block.

Functions

calculate_code_block_token_count

Return the <seen amount of consecutive fenced code-block tokens> + 1 that occur within a fenced code-block events.

cmark

As cmark_with_options(), but with default Options.

cmark_resume

As cmark_resume_with_options(), but with default Options.

cmark_resume_with_options

Serialize a stream of pulldown-cmark-Events into a string-backed buffer.

cmark_with_options

As cmark_resume_with_options(), but with the State finalized.